qmlui: more improvements to stage wizard
